Cleanup and tests
