+++ package.json
@@ -136,7 +136,6 @@
     "moment": "2.29.1",
     "moment-timezone": "0.5.33",
     "morgan": "1.10.0",
-    "node-vibrant": "^3.1.6",
     "path-to-regexp": "6.2.0",
     "polyfill-crypto.getrandomvalues": "1.0.0",
     "ramda": "0.27.1",
@@ -247,7 +246,6 @@
     "webpack-dev-server": "4.4.0",
     "webpack-merge": "5.8.0",
     "webpack-node-externals": "3.0.0",
-    "worker-loader": "^3.0.8",
     "yaml-js": "0.3.1"
   },
   "spec": {
